model/facial-beauty-prediction/requirements.txt
scipy==1.1.0
gevent==20.9.0
Keras==2.2.5
tensorflow~=1.15.3
numpy==1.16.1
dlib==19.16.0
Flask==1.1.2
opencv_python==3.4.5.20
Flask_Uploads==0.2.1

opencv-python~=3.4.5.20
pillow~=7.2.0
openpyxl~=3.0.5
werkzeug~=1.0.1
torch~=1.5.1
torchvision~=0.6.1
requests~=2.23.0
lxml~=4.5.2
tqdm~=4.46.0